Updated Simulation - has problems
The simulation has been written for some constant torque pulses. There's
a problem with the euler angles. The body-rates seem fine, but the euler
angles are blowing up. There's also an inherent issue of the
singularities at theta = n*pi.  Maybe that's it, but I may have
something wrong with the rotational kinematics. I'll work on a fix for
the next commit.

I also cleaned up the repo to only contain relevant or close-to-relevant
code.