Updated Controls to the VESCON situation
Not quite there yet. But the code is running fine without errors. Will
need to make sure that the controls still work with the new situation.
They were working with some simple changes (updating the torque
magnitude and changing the system pulse time).

After a dew more changes the code stipped torquing. Should not be a hard
problem to track down.